Gráfico do número de casos no Brasil ao longo do tempo para os dados do Brasil.io

Esta função plota o crescimento no número de casos no Brasil ao longo do tempo. Há duas opções de gráfico, veja o argumento `tipo` para mais detalhes.

plot_corona_br(df, log = TRUE, tipo = "numero")

Arguments

df
Data frame contendo o resultado da busca de `get_corona_br()`

log
Lógico. Se quer manter a escala log no eixo y do gráfico. Padrão log = TRUE. Apenas para `tipo = "numero"`

tipo
Caractere. Padrão `tipo = "numero"` para o número de casos ao longo do tempo. Usar `tipo = "aumento"` para plotar o aumento diário no número de casos