Extrai dados de coronavírus para o Brasil de Brasil.io

Esta função extrai os valores compilados pelo portal Brasil.io, que recolhe boletins informativos e casos do coronavírus por minicípio e por dia (disponível em: https://brasil.io/dataset/covid19/caso). A função salva o resultado no disco e escreve um arquivo com os metadados da requisição (metadado_corona_br.csv).

get_corona_br(
  dir = "output",
  filename = "corona_brasil",
  cidade = NULL,
  uf = NULL,
  ibge_cod = NULL,
  by_uf = FALSE
)

Arguments

dir
Diretório onde salvar o arquivo

filename
Nome do arquivo

cidade
Caractere indicando o(s) nome(s) do(s) município(s) brasileiro(s). Atenção, fornecer também o estado ou o código ibge para evitar ambiguidade. "Importados" extrai os casos importados, se disponíveis.

uf
Caractere indicando a abreviação do(s) estado(s) brasileiro(s)

ibge_cod
Numérico ou caractere. Código ibge do(s) município(s) ou estado(s) brasileiro(s). O código dos municípios tem 7 cifras e o código dos estados 2

by_uf
Lógico. Padrão by_uf = FALSE. Usar by_uf = TRUE se quiser os dados apenas por UF independente do município. Usar apenas quando não fornecer `cidade` ou `ibge_cod` de um município